+++ cmd/mailchain/commands/serve.go
@@ -18,8 +18,8 @@ import (
 	"fmt"
 
 	"github.com/mailchain/mailchain/cmd/mailchain/config"
-	"github.com/mailchain/mailchain/cmd/mailchain/internal/commands/prerun"
 	"github.com/mailchain/mailchain/cmd/mailchain/internal/http"
+	"github.com/mailchain/mailchain/cmd/mailchain/internal/prerun"
 	"github.com/spf13/cobra"
 	"github.com/ttacon/chalk"
 )
